refactor(pyproject): harden package metadata
