+++ .python-version
@@ -1,1 +1,1 @@
-3.11.2
+3.11
