+++ pyproject.toml
@@ -1,13 +1,18 @@
 [build-system]
-requires = ["hatchling>=1.25"]
+requires = ["hatchling>=1.27"]
 build-backend = "hatchling.build"
 
 [project]
-name = "ecli"
+name = "ecli-editor"
 version = "0.1.0"
 description = "ECLI — fast terminal code editor"
 readme = "README.md"
 requires-python = ">=3.11"
+license = "Apache-2.0"
+license-files = ["LICENSE*"]
+authors = [
+    { name = "Siergej Sobolewski", email = "s.sobolewski@hotmail.com" },
+]
 keywords = [
     "editor",
     "terminal",
@@ -16,19 +21,15 @@ keywords = [
     "text-editor",
     "developer-tools",
     "devops",
-    "ecli",
-    "IDE",
 ]
-
 classifiers = [
     "Development Status :: 4 - Beta",
     "Environment :: Console :: Curses",
     "Intended Audience :: Developers",
-    "License :: OSI Approved :: Apache Software License",
-    "Operating System :: POSIX",
-    "Operating System :: MacOS",
     "Operating System :: POSIX :: Linux",
     "Operating System :: POSIX :: BSD :: FreeBSD",
+    "Operating System :: MacOS",
+    "Operating System :: Microsoft :: Windows",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -37,40 +38,25 @@ classifiers = [
     "Topic :: Text Editors",
     "Typing :: Typed",
 ]
-
 dependencies = [
-    "Pygments>=2.19.1; sys_platform != 'win32'",
-    "pyperclip>=1.9.0; sys_platform != 'win32'",
-    "wcwidth>=0.2.13; sys_platform != 'win32'",
-    "chardet>=5.2.0; sys_platform != 'win32'",
-    "PyYAML>=6.0.2; sys_platform != 'win32'",
-    "toml>=0.10.2; sys_platform != 'win32'",
-    "attrs>=24.0.0; sys_platform != 'win32'",
-    "typing_extensions>=4.12.2; sys_platform != 'win32'",
-    "packaging>=24.1; sys_platform != 'win32'",
-    "aiohttp>=3.9.0; sys_platform != 'win32'",
-    "pygls>=1.3.1; sys_platform != 'win32'",
-    "lsprotocol>=2023.0.1; sys_platform != 'win32'",
-    "libcst>=1.5.0; sys_platform != 'win32'",
-    "cattrs>=23.2.0; sys_platform != 'win32'",
-    "python-dotenv>=1.0.1; sys_platform != 'win32'",
-    "tato>=0.2.3; sys_platform != 'win32'",
+    "Pygments>=2.19.1",
+    "pyperclip>=1.9.0",
+    "wcwidth>=0.2.13",
+    "chardet>=5.2.0",
+    "PyYAML>=6.0.2",
+    "attrs>=24.0.0",
+    "typing_extensions>=4.12.2",
+    "packaging>=24.1",
+    "aiohttp>=3.9.0",
+    "pygls>=1.3.1",
+    "lsprotocol>=2023.0.1",
+    "libcst>=1.5.0",
+    "cattrs>=23.2.0",
+    "python-dotenv>=1.0.1",
+    "toml>=0.10.2",
+    "tato>=0.2.3",
+    "windows-curses>=2.4.0; sys_platform == 'win32'",
 ]
-
-[project.license]
-text = "Apache-2.0"
-
-[[project.authors]]
-name = "Siergej Sobolewski"
-email = "s.sobolewski@hotmail.com"
-
-[project.scripts]
-ecli = "ecli.main:start"
-
-[project.urls]
-Homepage = "https://ecli.io"
-Repository = "https://github.com/SSobol77/ecli"
-"Bug Tracker" = "https://github.com/SSobol77/ecli/issues"
 
 [project.optional-dependencies]
 dev = [
@@ -79,7 +65,6 @@ dev = [
     "mypy>=1.10.0",
     "mypy-extensions>=1.0.0",
     "types-PyYAML>=6.0.12.20240516",
-    "types-toml>=0.10.8.20240310",
     "types-pyperclip>=1.9.0.20220518",
     "pytest>=8.4.1",
     "pytest-asyncio>=0.23.8",
@@ -87,17 +72,50 @@ dev = [
     "pytest-mock>=3.14.0",
     "pytest-cov>=6.0.0",
     "coverage>=7.4.0",
+    "build>=1.2",
+    "twine>=5.0",
+]
+
+[project.scripts]
+ecli = "ecli.__main__:main"
+
+[project.urls]
+Homepage = "https://ecli.io"
+Documentation = "https://ecli.io/docs"
+Repository = "https://github.com/SSobol77/ecli"
+Issues = "https://github.com/SSobol77/ecli/issues"
+Changelog = "https://github.com/SSobol77/ecli/blob/main/CHANGELOG.md"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/ecli"]
+
+[tool.hatch.build.targets.sdist]
+include = [
+    "/src",
+    "/tests",
+    "/README.md",
+    "/LICENSE",
+    "/CHANGELOG.md",
+    "/pyproject.toml",
+    "/main.py",
 ]
 
 [tool.pytest.ini_options]
 minversion = "8.0"
-addopts = "-ra --tb=short --strict-markers"
+addopts = "-ra --tb=short --strict-markers --strict-config"
 testpaths = ["tests"]
 python_files = "test_*.py"
-markers = ["asyncio"]
+asyncio_mode = "auto"
+markers = [
+    "asyncio: marks async tests (auto-applied via asyncio_mode)",
+    "integration: integration tests requiring external resources",
+    "slow: tests that take >1 second",
+]
 
 [tool.mypy]
-files = ["ecli", "tests"]
+files = ["src/ecli", "tests"]
+mypy_path = "src"
+python_version = "3.11"
 ignore_missing_imports = true
 show_error_codes = true
 strict = true
@@ -105,14 +123,32 @@ warn_unused_ignores = true
 warn_return_any = true
 warn_unreachable = true
 no_implicit_optional = true
-python_version = "3.13"
-
 
 [tool.ruff]
 line-length = 88
 target-version = "py311"
-extend-exclude = ["__pycache__", "*.pyc", ".git"]
-
+extend-exclude = [
+    "__pycache__",
+    "*.pyc",
+    ".git",
+    "build",
+    "dist",
+    "src/ecli/core/Ecli.py",
+    "src/ecli/core/History.py",
+    "src/ecli/integrations/AI.py",
+    "src/ecli/integrations/GitBridge.py",
+    "src/ecli/integrations/LinterBridge.py",
+    "src/ecli/integrations/__init__.py",
+    "src/ecli/ui/DrawScreen.py",
+    "src/ecli/ui/KeyBinder.py",
+    "src/ecli/ui/PanelManager.py",
+    "src/ecli/ui/TerminalAppMode.py",
+    "src/ecli/ui/__init__.py",
+    "src/ecli/ui/panels.py",
+    "src/ecli/utils/__init__.py",
+    "src/ecli/utils/logging_config.py",
+    "src/ecli/utils/utils.py",
+]
 
 [tool.ruff.lint]
 select = ["E", "F", "W", "C90", "I", "UP", "B", "N", "Q", "PL", "RET", "D"]
@@ -131,45 +167,25 @@ ignore = [
     "UP035",
 ]
 
-
 [tool.ruff.lint.per-file-ignores]
 "tests/**/*.py" = ["D100", "D101", "D102", "D103", "D104", "F401", "E402"]
 "src/ecli/__init__.py" = ["F401"]
-
+"src/ecli/__main__.py" = ["C901", "D100", "D212", "I001"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "google"
 
-
 [tool.ruff.lint.isort]
 combine-as-imports = true
 lines-after-imports = 2
-extra-standard-library = ["typing_extensions"]
-
+extra-standard-library = ["typing_extensions", "tomllib"]
 
 [tool.ruff.lint.flake8-quotes]
 docstring-quotes = "double"
 inline-quotes = "double"
-
 
 [tool.ruff.format]
 quote-style = "double"
 indent-style = "space"
 skip-magic-trailing-comma = false
 line-ending = "lf"
-
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/ecli"]
-include = ["ecli/py.typed"]
-
-
-[tool.hatch.build.targets.sdist]
-include = [
-    "/ecli",
-    "/tests",
-    "/README.md",
-    "/LICENSE",
-    "/CHANGELOG.md",
-    "/pyproject.toml",
-]
